Fix project components after split
