+++ src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Navbar from './components/Navbar.jsx'
 import Footer from './components/Footer.jsx'
 import Home from './components/Home.jsx'
-import ProjectDetails from './components/ProjectDetails.jsx'
+import ProjectDetails from './components/projects/ProjectDetails.jsx'
 import './App.css'
 
 function App() {
