+++ src/components/Home.jsx
@@ -1,6 +1,6 @@
 import Hero from './Hero.jsx'
 import About from './About.jsx'
-import Projects from './Projects.jsx'
+import Projects from './projects/Projects.jsx'
 import Blog from './Blog.jsx'
 import { sections } from '../config.js'
 import useScrollAnimation from '../hooks/useScrollAnimation.js'
