+++ src/components/projects/Projects.jsx
@@ -1,9 +1,7 @@
-import { Link, useNavigate } from 'react-router-dom'
 import { projectList } from '../../projects'
-import { RenderCard } from 'src/components/projects/projectCardRenderer.js'
+import RenderCard from './projectCardRenderer.js'
 
 function Projects() {
-  const navigate = useNavigate()
 
   const commercial = projectList.filter(p => p.category === 'commercial')
   const home = projectList.filter(p => p.category === 'home')
@@ -12,9 +10,17 @@ function Projects() {
     <section id="projects" className="section">
       <div className="container">
         <h3 className="section-title scroll-animate">Commercial Projects</h3>
-        <div className="projects-grid">{commercial.map(renderCard)}</div>
+        <div className="projects-grid">
+          {commercial.map(p => (
+            <RenderCard key={p.id} project={p} />
+          ))}
+        </div>
         <h3 className="section-title scroll-animate">Home Projects</h3>
-        <div className="projects-grid">{home.map(renderCard)}</div>
+        <div className="projects-grid">
+          {home.map(p => (
+            <RenderCard key={p.id} project={p} />
+          ))}
+        </div>
       </div>
     </section>
   )
